fix(import-cycles): break contracts<->project and cli<->gateway new cycles (astrid fixer import-cycle-contracts-project-20260815T1250Z)
